Display horizontal distance in km/h or m/h
The Horus team found that horizontal distances in m/s were hard to get
an intuition for, particularly as the speed across land is often
used for calculating if a car can travel to the landing site in time.

This change modifies the horizontal speed to be calculated in km/h, or
m/h if the user has selected imperial units.

Signed-off-by: Joel Stanley <joel@jms.id.au>